Claim or yield asset dynamically
On a /asset/info page, claim or yield the asset with the claim or
yield button.

Under the hood, this is a PUT request, followed by a POST request to
redirect. We could optimize this with AJAX, but my pages are relatively
lightweight & I dont want more complexity for this simple use demo.

TODO: Bugfix / QOL updates for visuals?